backend/requirements.txt
alembic==1.14.1
blinker==1.9.0
click==8.1.8
colorlog==6.9.0
Flask==3.1.0
flask-cors==5.0.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.6.1
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.6.0
packaging==24.2
psycopg2==2.9.10
python-dotenv==1.0.1
sqlacodegen==2.3.0.post1
SQLAlchemy==2.0.38
typeguard==4.4.2
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
